Updates to AddAnonymityRevoker/AddIdentityProvider based on feedback
